chore: tidies namespaces
