+++ Editor/UI/Windows/ReadyPlayerMeEditor.cs
@@ -15,7 +15,7 @@ namespace ReadyPlayerMe.Editor.UI.Windows
         private DeveloperLoginView _developerLoginView;
         private ApplicationManagementView _applicationManagementView;
 
-        [MenuItem("Tools/Ready Player Me")]
+        [MenuItem("Tools/Ready Player Me/Project Settings")]
         public static void Generate()
         {
             var window = GetWindow<ReadyPlayerMeEditor>("Ready Player Me");
