+++ Runtime/Data/ScriptableObjects/AvatarSkeletonDefinition.cs.meta
