+++ Runtime/Data/V1/Avatar.cs
