--- Runtime/Scripts/Data/ScriptableObjects/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 63ac0a2246843b64ba17621bc57f58b4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
